src/integration-test-agent/pyproject.toml
[tool.poetry]
name = "integration-test-agent"
version = "1.0.0"
description = "ORBIT Integration Test Agent — Postman collection and Newman config generator"
authors = ["AdPorts AI Platform Team"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
httpx = "^0.27.0"
langgraph = "^0.2.28"
langchain-core = "^0.3.15"
litellm = "^1.50.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
